% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.Pe.R
\docType{data}
\name{datos.Pe}
\alias{datos.Pe}
\title{Datos historicos de estaciones del Peru (SENAMHI).}
\format{
un marco de datos con 22 series de tiempo mensuales. La primera fila
corresponde a la fecha, las siguientes corresponden a las series de tiempo mensuales.
\describe{
  \item{Fecha}{el registro de fechas es mensual, desde 01-01-1964 hasta 31-12-2016}
  ...
}
}
\usage{
datos.Pe
}
\description{
Un conjunto de datos de 22 series históricas de estaciones del Perú (SENAMHI).
}
\keyword{datasets}
