% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estaciones.Pe.R
\docType{data}
\name{estaciones.Pe}
\alias{estaciones.Pe}
\title{Estaciones del Peru (SENAMHI).}
\format{
un marco de datos con 22 filas de información de estaciones.
\describe{
  \item{Estacion}{Nombre de la estación}
  \item{Lat}{Latitud}
  \item{Lon}{Longitud}
  \item{Alt}{Altitud, en msnm.}
  ...
}
}
\usage{
estaciones.Pe
}
\description{
Un conjunto de datos de 22 estaciones del Perú (SENAMHI)
.
}
\keyword{datasets}
