% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his.aed.R
\name{his.aed}
\alias{his.aed}
\title{Funcion para obtener el histograma de datos mensuales para AED.}
\usage{
his.aed(data, estaciones, i = NULL, col = NULL, variable = NULL, um = NULL)
}
\arguments{
\item{data}{conjunto de datos (dataframe) de series mensuales.}

\item{estaciones}{conjunto de datos (dataframe) de las estaciones.}

\item{i}{N째 de orden (integer) de la estaci처n}

\item{col}{color de los datos dudosos (outliers). --}

\item{variable}{nombre de la variable hidrol처gica.}

\item{um}{unidad de medida. ejm.: mm (Precipitaci처n).}
}
\description{
Funcion para obtener el histograma de datos mensuales para AED.
}
