% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infhis.R
\name{infhis}
\alias{infhis}
\title{Funcion para obtener un resumen del histograma de datos mensuales para AED.}
\usage{
infhis(data, estaciones)
}
\arguments{
\item{data}{conjunto de datos (dataframe) de series mensuales.}

\item{estaciones}{conjunto de datos (dataframe) de las estaciones.}
}
\description{
Funcion para obtener un resumen del histograma de datos mensuales para AED.
}
