% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infna.R
\name{infna}
\alias{infna}
\title{Información no disponible}
\usage{
infna(data, estaciones)
}
\arguments{
\item{data}{conjunto de datos (dataframe) de series mensuales.}

\item{estaciones}{conjunto de datos (dataframe) de las estaciones.}
}
\description{
Información no disponible
}
