% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infout.R
\name{infout}
\alias{infout}
\title{Funcion para obtener la cantidad, valor maximo y minimo de datos dudosos.}
\usage{
infout(data, estaciones)
}
\arguments{
\item{data}{conjunto de datos (dataframe) de series mensuales.}

\item{estaciones}{conjunto de datos (dataframe) de las estaciones.}
}
\description{
Funcion para obtener la cantidad, valor maximo y minimo de datos dudosos.
}
