% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest}
\alias{rest}
\title{Funcion para obtener un resumen estadistico de un conjunto de series de tiempo.}
\usage{
rest(data, estaciones)
}
\arguments{
\item{data}{conjunto de datos (dataframe) de series mensuales.}

\item{estaciones}{conjunto de datos (dataframe) de las estaciones.}
}
\description{
Funcion para obtener un resumen estadistico de un conjunto de series de tiempo.
}
